added export PDF + routing feature
