+++ package.json
@@ -19,6 +19,8 @@
     "@types/react-dom": "^18.0.5",
     "debounce": "^1.2.1",
     "framer-motion": "^6.3.16",
+    "html2canvas": "^1.4.1",
+    "jspdf": "^2.5.1",
     "re-resizable": "^6.9.9",
     "react": "^18.2.0",
     "react-dnd": "^16.0.1",
@@ -26,6 +28,7 @@
     "react-rectangle-selection": "^1.0.4",
     "react-redux": "^8.0.2",
     "react-rnd": "^10.3.7",
+    "react-router-dom": "^6.4.0",
     "react-scripts": "5.0.1",
     "redux": "^4.2.0",
     "typescript": "^4.7.4",
@@ -33,8 +36,8 @@
   },
   "scripts": {
     "start": "react-scripts start",
-    "predeploy" : "npm run build",
-    "deploy" : "gh-pages -d build",
+    "predeploy": "npm run build",
+    "deploy": "gh-pages -d build",
     "build": "react-scripts build",
     "test": "react-scripts test",
     "eject": "react-scripts eject",
